zkir-v3: remove 'decode' instruction
