+++ zkir-v3/src/ir_instructions/mod.rs
@@ -18,7 +18,6 @@ type F = outer::Scalar;
 pub mod add;
 pub mod assign;
 pub mod constrain_eq;
-pub mod decode;
 pub mod ec_mul;
 pub mod encode;
 pub mod eq;
