+++ zkir-v3/tests/proofs.rs
@@ -1029,54 +1029,33 @@ mod proof_tests {
     #[actix_rt::test]
     async fn test_secp256k1_proof() {
         // Single circuit exercising all three Secp256k1 types.
-        // Base and Scalar values are decoded from native limbs, added, then
-        // encoded back; a round-trip decode verifies encode is its inverse.
-        // Points are typed inputs; their sum is checked via a private input.
+        // Base, Scalar and Point values are typed inputs; arithmetic results
+        // are checked via private inputs.
         let ir_raw = r#"{
            "version": { "major": 3, "minor": 0 },
            "inputs": [
-              { "name": "%id",   "type": "Point<Secp256k1>"  },
-              { "name": "%p0",   "type": "Point<Secp256k1>"  },
-              { "name": "%p1",   "type": "Point<Secp256k1>"  },
-              { "name": "%b0_0", "type": "Scalar<BLS12-381>" },
-              { "name": "%b0_1", "type": "Scalar<BLS12-381>" },
-              { "name": "%b0_2", "type": "Scalar<BLS12-381>" },
-              { "name": "%b0_3", "type": "Scalar<BLS12-381>" },
-              { "name": "%b1_0", "type": "Scalar<BLS12-381>" },
-              { "name": "%b1_1", "type": "Scalar<BLS12-381>" },
-              { "name": "%b1_2", "type": "Scalar<BLS12-381>" },
-              { "name": "%b1_3", "type": "Scalar<BLS12-381>" },
-              { "name": "%s0_0", "type": "Scalar<BLS12-381>" },
-              { "name": "%s0_1", "type": "Scalar<BLS12-381>" },
-              { "name": "%s0_2", "type": "Scalar<BLS12-381>" },
-              { "name": "%s0_3", "type": "Scalar<BLS12-381>" },
-              { "name": "%s1_0", "type": "Scalar<BLS12-381>" },
-              { "name": "%s1_1", "type": "Scalar<BLS12-381>" },
-              { "name": "%s1_2", "type": "Scalar<BLS12-381>" },
-              { "name": "%s1_3", "type": "Scalar<BLS12-381>" }
+              { "name": "%id", "type": "Point<Secp256k1>"  },
+              { "name": "%p0", "type": "Point<Secp256k1>"  },
+              { "name": "%p1", "type": "Point<Secp256k1>"  },
+              { "name": "%b0", "type": "Base<Secp256k1>"   },
+              { "name": "%b1", "type": "Base<Secp256k1>"   },
+              { "name": "%s0", "type": "Scalar<Secp256k1>" },
+              { "name": "%s1", "type": "Scalar<Secp256k1>" }
            ],
            "outputs": [],
            "do_communications_commitment": false,
            "instructions": [
-               { "op": "decode", "type": "Base<Secp256k1>",   "inputs": ["%b0_0","%b0_1","%b0_2","%b0_3"], "output": "%b0" },
-               { "op": "decode", "type": "Base<Secp256k1>",   "inputs": ["%b1_0","%b1_1","%b1_2","%b1_3"], "output": "%b1" },
-               { "op": "decode", "type": "Scalar<Secp256k1>", "inputs": ["%s0_0","%s0_1","%s0_2","%s0_3"], "output": "%s0" },
-               { "op": "decode", "type": "Scalar<Secp256k1>", "inputs": ["%s1_0","%s1_1","%s1_2","%s1_3"], "output": "%s1" },
                { "op": "add", "a": "%p0", "b": "%p1", "output": "%p2" },
                { "op": "add", "a": "%b0", "b": "%b1", "output": "%b2" },
                { "op": "add", "a": "%s0", "b": "%s1", "output": "%s2" },
                { "op": "mul", "a": "%b0", "b": "%b1", "output": "%b_prod" },
                { "op": "mul", "a": "%s0", "b": "%s1", "output": "%s_prod" },
-               { "op": "encode", "input": "%b2", "outputs": ["%b2_0","%b2_1","%b2_2","%b2_3"] },
-               { "op": "encode", "input": "%s2", "outputs": ["%s2_0","%s2_1","%s2_2","%s2_3"] },
-               { "op": "decode", "type": "Base<Secp256k1>",   "inputs": ["%b2_0","%b2_1","%b2_2","%b2_3"], "output": "%b2_rt" },
-               { "op": "decode", "type": "Scalar<Secp256k1>", "inputs": ["%s2_0","%s2_1","%s2_2","%s2_3"], "output": "%s2_rt" },
-               { "op": "neg", "a": "%p0",  "output": "%p0_neg" },
-               { "op": "neg", "a": "%b0",  "output": "%b0_neg" },
-               { "op": "neg", "a": "%s0",  "output": "%s0_neg" },
-               { "op": "inv", "a": "%b0",  "output": "%b0_inv" },
-               { "op": "inv", "a": "%s0",  "output": "%s0_inv" },
-               { "op": "private_input", "type": "Point<Secp256k1>",  "guard": null, "output": "%p2_priv"    },
+               { "op": "neg", "a": "%p0", "output": "%p0_neg" },
+               { "op": "neg", "a": "%b0", "output": "%b0_neg" },
+               { "op": "neg", "a": "%s0", "output": "%s0_neg" },
+               { "op": "inv", "a": "%b0", "output": "%b0_inv" },
+               { "op": "inv", "a": "%s0", "output": "%s0_inv" },
+               { "op": "private_input", "type": "Point<Secp256k1>",  "guard": null, "output": "%p2_priv"     },
                { "op": "private_input", "type": "Base<Secp256k1>",   "guard": null, "output": "%b_prod_priv" },
                { "op": "private_input", "type": "Scalar<Secp256k1>", "guard": null, "output": "%s_prod_priv" },
                { "op": "private_input", "type": "Point<Secp256k1>",  "guard": null, "output": "%p0_neg_priv" },
@@ -1085,12 +1064,9 @@ mod proof_tests {
                { "op": "private_input", "type": "Base<Secp256k1>",   "guard": null, "output": "%b0_inv_priv" },
                { "op": "private_input", "type": "Scalar<Secp256k1>", "guard": null, "output": "%s0_inv_priv" },
                { "op": "constrain_eq", "a": "%p2",     "b": "%p2_priv"     },
-               { "op": "constrain_eq", "a": "%b2",     "b": "%b2_rt"       },
                { "op": "constrain_eq", "a": "%b_prod", "b": "%b_prod_priv" },
                { "op": "constrain_eq", "a": "%p0_neg", "b": "%p0_neg_priv" },
                { "op": "constrain_eq", "a": "%b0_neg", "b": "%b0_neg_priv" },
-               { "op": "test_eq",      "a": "%s2",     "b": "%s2_rt",    "output": "%s_eq"   },
-               { "op": "assert",       "cond": "%s_eq" },
                { "op": "test_eq",      "a": "%s_prod", "b": "%s_prod_priv", "output": "%sp_eq" },
                { "op": "assert",       "cond": "%sp_eq" },
                { "op": "test_eq",      "a": "%s0_neg", "b": "%s0_neg_priv", "output": "%sn_eq" },
@@ -1117,8 +1093,6 @@ mod proof_tests {
                 .collect()
         };
 
-        // p0, p1 are typed Point<Secp256k1> inputs (8 limbs each);
-        // b0, b1, s0, s1 are passed as raw native limbs (4 each) for decode.
         let inputs: Vec<transient_crypto::curve::Fr> = [
             encode(IrValue::Secp256k1Point(id)),
             encode(IrValue::Secp256k1Point(p0)),
